overhaul dynamo to make it more modular
