+++ dynamo/external/hodge.py
@@ -8,8 +8,8 @@ from scipy.sparse import csr_matrix
 from scipy.linalg import qr
 from itertools import combinations
 from igraph import Graph
-from ..tools.scVectorField import graphize_vecfld
-from ..tools.utils_vecCalc import vecfld_from_adata, vector_field_function
+from ..vectorfield.scVectorField import graphize_vecfld
+from ..vectorfield.utils_vecCalc import vecfld_from_adata, vector_field_function
 from ..tools.sampling import trn, sample_by_velocity
 
 def gradop(g):
