+++ dynamo/movie/__init__.py
@@ -1,0 +1,4 @@
+"""Mapping Vector Field of Single Cells
+"""
+
+from .fate import StreamFuncAnim, animate_fates
