+++ dynamo/plot/dynamics.py
@@ -7,8 +7,8 @@ from .utils import arrowed_spines, despline_all, deaxis_all
 from .utils import quiver_autoscaler, default_quiver_args
 from .utils import save_fig
 from .scatters import scatters
-from ..tools.velocity import sol_u, sol_s, solve_first_order_deg
-from ..tools.utils_moments import moments
+from ..estimation.csc.velocity import sol_u, sol_s, solve_first_order_deg
+from ..estimation.tsc.utils_moments import moments
 from ..tools.utils import get_mapper, log1p_
 from ..tools.utils import update_dict, get_valid_bools
 from ..configuration import _themes
