+++ dynamo/tools/time_series.py
@@ -1,7 +1,7 @@
 import numpy as np
 from scipy.sparse import csr_matrix
 from scipy.sparse.csgraph import minimum_spanning_tree
-from .DDRTree import DDRTree_py
+from .DDRTree_py import DDRTree_py
 
 
 def cal_ncenter(ncells, ncells_limit=100):
