+++ dynamo/vectorfield/Bhattacharya.py
