+++ dynamo/vectorfield/Wang.py
