+++ dynamo/vectorfield/__init__.py
@@ -1,0 +1,46 @@
+"""Mapping Vector Field of Single Cells
+"""
+
+from .scVectorField import (
+    SparseVFC,
+    graphize_vecfld,
+    vectorfield
+)  # , evaluate, con_K_div_cur_free
+from .utils_vecCalc import (
+    vector_field_function,
+)
+from .topography import FixedPoints, VectorField2D, topography, VectorField
+from .cell_accelerations import cell_accelerations
+from .vector_calculus import (
+    speed,
+    jacobian,
+    curl,
+    divergence,
+    acceleration,
+    curvature,
+    torsion,
+    rank_speed_genes,
+    rank_divergence_genes,
+    rank_acceleration_genes,
+    rank_curvature_genes,
+)
+
+# potential related
+from .scPotential import (
+    search_fixed_points,
+    gen_fixed_points,
+    gen_gradient,
+    IntGrad,
+    DiffusionMatrix,
+    action,
+    Potential,
+)  # , vector_field_function
+from .Bhattacharya import path_integral, alignment
+from .Wang import Wang_action, Wang_LAP, transition_rate, MFPT
+from .Ao import Ao_pot_map, solveQ
+
+# stochastic process related
+from .stochastic_process import diffusionMatrix
+
+# vfGraph operation related:
+from .vfGraph import vfGraph
