+++ dynamo/vectorfield/utils.py
@@ -1,0 +1,6 @@
+import numpy as np
+
+def is_outside_domain(x, domain):
+    x = x[None, :] if x.ndim == 1 else x
+    return np.any(np.logical_or(x < domain[0], x > domain[1]), axis=1)
+
